requirements.txt
asgiref==3.7.2
Django==5.0.1
django-bootstrap5==23.4
django-crispy-forms==2.1
django-filter==23.5
django-tables2==2.7.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pytest==7.4.4
PyYAML==6.0.1
sqlparse==0.4.4
